Remove confusing paragraph in README regarding post redirects
As it's now, the paragraph points to an "example" without making clear 
where the reader can find such example nor the function referenced.

I could have tried to fix that and link to the example projects. But I see
little point on doing that as the previous line (which talks about POSTing
or sending the key information to the backend) already documents the
expected behavior.