Document the versioning options in README
Also change the default to 1.x.x (auto-update as long as changes are backwards compatible). 

I tried to explain why the default would be 1.5.x and the options would be to fix it to 1.5.10 (stricter) 
or 1.x.x (simpler) and found out that as general advice it's hard to justify. It's much easier to explain 
that the default is the simpler way (and we take responsibility for not breaking things in new minor 
releases!) and that people can be more strict if they want to be more conservative/strict.